move doc examples from DatePicker to Date
