+++ packages/dnb-design-system-portal/src/docs/uilib/components/date-picker/demos.mdx
@@ -99,12 +99,6 @@ import enUS from '@dnb/eufemia/shared/locales/en-US'
 
 <DatePickerErrorStatus />
 
-### DatePicker with date limit validation
-
-The DatePicker will automatically display an error message if the selected date or dates are outside the given date limits.
-
-<DatePickerDateLimitValidation />
-
 ### Opened DatePicker
 
 <DatePickerCalendar />
