+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/forms/feature-fields/Date/demos.mdx
@@ -38,6 +38,12 @@ To enable the picker to close automatically, you have to set `showCancelButton` 
 
 <Examples.WithError />
 
+### Date limit validation
+
+The Date field will automatically display an error message if the selected date or dates are outside the given date limits.
+
+<Examples.DatePickerDateLimitValidation />
+
 ### Validation - Required
 
 <Examples.ValidationRequired />
